+++ pkg/operator/test/deployment.go
@@ -1,0 +1,64 @@
+/*
+Copyright 2018 The Rook Authors. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package test
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/rook/rook/pkg/operator/k8sutil"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+)
+
+// WaitForDeploymentImage waits for all deployments with the given labels are running.
+func WaitForDeploymentImage(clientset kubernetes.Interface, namespace, label, container string, initContainer bool, desiredImage string) error {
+
+	sleepTime := 3
+	attempts := 120
+	for i := 0; i < attempts; i++ {
+		deployments, err := clientset.Apps().Deployments(namespace).List(metav1.ListOptions{LabelSelector: label})
+		if err != nil {
+			return fmt.Errorf("failed to list deployments with label %s. %v", label, err)
+		}
+
+		matches := 0
+		for _, d := range deployments.Items {
+			image, err := k8sutil.GetDeploymentSpecImage(clientset, d, container, initContainer)
+			if err != nil {
+				logger.Infof("failed to get image for deployment %s. %+v", d.Name, err)
+				continue
+			}
+			if image == desiredImage {
+				matches++
+			}
+		}
+
+		if matches == len(deployments.Items) && matches > 0 {
+			logger.Infof("all %d %s deployments are on image %s", matches, label, desiredImage)
+			return nil
+		}
+
+		if len(deployments.Items) == 0 {
+			logger.Infof("waiting for at least one deployment to start to see the version")
+		} else {
+			logger.Infof("%d/%d %s deployments match image %s", matches, len(deployments.Items), label, desiredImage)
+		}
+		time.Sleep(time.Duration(sleepTime) * time.Second)
+	}
+	return fmt.Errorf("failed to wait for image %s in label %s", desiredImage, label)
+}
